refactor(hashers): reorg
